requirements.txt
alembic==1.17.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
bcrypt==4.0.1
blis==1.3.3
catalogue==2.0.10
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudinary==1.44.1
cloudpathlib==0.23.0
confection==0.1.5
cryptography==46.0.3
cymem==2.0.13
Deprecated==1.3.1
deprecation==2.1.0
distro==1.9.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
fastapi==0.122.0
fastapi-cli==0.0.16
fastapi-cloud-cli==0.5.2
fastar==0.8.0
greenlet==3.2.4
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.3
limits==5.6.0
lxml==6.0.2
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
multidict==6.7.0
murmurhash==1.0.15
nltk==3.9.2
numpy==2.4.0
openai==2.15.0
packaging==25.0
passlib==1.7.4
pdfminer.six==20251107
pdfplumber==0.11.8
phonenumbers==9.0.21
pillow==12.0.0
postgrest==2.24.0
preshed==3.0.12
propcache==0.4.1
psycopg==3.2.13
psycopg-binary==3.2.13
psycopg2-binary==2.9.11
pyasn1==0.6.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
PyMuPDF==1.23.9
PyMuPDFb==1.23.9
pypdfium2==5.2.0
python-docx==1.1.0
python-dotenv==1.2.1
python-jose==3.3.0
python-magic==0.4.27
python-multipart==0.0.20
PyYAML==6.0.3
realtime==2.24.0
regex==2025.11.3
requests==2.32.5
resend==2.19.0
rich==14.2.0
rich-toolkit==0.17.0
rignore==0.7.6
rsa==4.9.1
sentry-sdk==2.46.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
slowapi==0.1.9
smart_open==7.5.0
sniffio==1.3.1
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.44
srsly==2.5.2
starlette==0.50.0
storage3==2.24.0
StrEnum==0.4.15
supabase==2.24.0
supabase-auth==2.24.0
supabase-functions==2.24.0
thinc==8.3.10
tqdm==4.67.1
typer==0.20.0
typer-slim==0.21.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
wasabi==1.1.3
watchfiles==1.1.1
weasel==0.4.3
websockets==15.0.1
wrapt==2.0.1
yarl==1.22.0
google-auth==2.28.0                    # ← ADD
google-auth-oauthlib==1.2.0            # ← ADD
google-auth-httplib2==0.2.0          # ← ADD
apscheduler==3.10.4